requirements.txt
tensorflow-gpu==2.1.0
numpy==1.18.1
timebudget==0.7